requirements.txt
captcha
chromadb
discord.py
pydantic
python-dotenv
streamlit
tinydb
together